requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Supabase integration
supabase==2.18.1
python-jose[cryptography]==3.3.0
websockets>=12.0

# Document processing
PyMuPDF==1.26.4
python-docx==1.2.0

# AI/OpenAI integration
openai==1.104.2

# Type hints and utilities
typing_extensions==4.14.0
python-dateutil==2.9.0.post0
pydantic==2.11.7
pydantic-settings==2.6.1

# HTTP client for external requests
httpx>=0.26.0,<0.29.0
requests>=2.31.0

# JSON
json5

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
